fix: address Gemini review findings on PR #131
- Extract magic numbers to src/lib/invites/constants.ts:
  INVITE_WEEKLY_LIMIT=3, INVITE_EXPIRY_DAYS=7, INVITE_CODE_MIN_LENGTH=4,
  REFERRAL_TREE_MAX_DEPTH=3, APP_URL
- Extract nextMondayUTC() to src/lib/invites/next-monday.ts
  (was duplicated in invites/route.ts AND invites/page.tsx)
- invites/route.ts and invites/page.tsx now import from shared modules
- Migration: document configurable values with cross-references to
  constants.ts so they stay in sync
- No more hardcoded APP_URL strings in invite routes
